requirements.txt
# HomeDome Content Engine - Dependencies
# Core
google-generativeai>=0.8.0
elevenlabs>=1.0.0
Pillow>=10.0.0
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Post-Bridge upload (simple REST API)
requests>=2.31.0

# Video processing
mutagen>=1.47.0

# Subtitles
pysrt>=1.1.2

# Utilities
Jinja2>=3.1.0
click>=8.1.0
rich>=13.0.0
